addresses PR comments
