+++ backstory/src/main/java/com/google/sps/data/testImages/snowMountain.JPG
